# chore: betas optimization - 2^d multiplications