With 16 cores for the avm trace (2^21), the time goes from ca. 80 ms to 37 ms.
For a trace of size 2^24, timing goes from 600 ms to 280ms.

Relevant issue [#864](https://github.com/AztecProtocol/barretenberg/issues/864)